% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neural.R
\name{random_ann}
\alias{random_ann}
\title{Creates a random function with neural netowrk structure}
\usage{
random_ann(nunits, vars = 0 * nunits + 1, sparsities = nunits,
  transfer = c("logistic", "ppart", "sign"))
}
\arguments{
\item{nunits}{A vector indicating the number of layers in each unit, the first element being the input dimension
NOTE: Omit the 1-unit output layer}

\item{vars}{Marginal variance of coefficients for each layer, default = 1}

\item{sparsities}{Sparsity of coefficients for each layer, defaul = ndim}

\item{transfer}{Choice of transfer function}
}
\description{
This function returns a closure with given input dimensionality,
using a neural network with random weights
}
\examples{
nunits <- c(5, 10)
x <- matrix(rnorm(100 * 5), 100, 5)
f <- random_ann(nunits)
y <- f(x)
}
\keyword{network,}
\keyword{neural}
\keyword{perceptron}

