% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/basic.R
\name{random_linear}
\alias{random_linear}
\title{Creates a random linear function}
\usage{
random_linear(ndim, var = 1, sparsity = ndim, positive = FALSE)
}
\arguments{
\item{ndim}{Dimension of input (x)}

\item{var}{Marginal variance of coefficients, default = 1}

\item{sparsity}{Sparsity of coefficients, defaul = ndim}

\item{positive}{Constrain the coefficients to positive? default = False}
}
\description{
This function returns a closure with given input dimensionality and coefficients
}
\examples{
f <- random_linear(10, 1, 2)
f(1:10)
f()
}
\keyword{linear}

