% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/neural.R
\name{transfer_functions}
\alias{transfer_functions}
\title{Transfer functions for neural networks}
\usage{
transfer_functions(transfer = c("logistic", "ppart", "sign"))
}
\arguments{
\item{tranfer}{A string indicating the name of the transfer function}
}
\description{
Thus function returns a specified transfer function
}

